(Prelude) Factor out function updateFootnotes
